requirements.txt
pandas==3.0.1
openpyxl==3.1.5
xlrd==2.0.2
python-docx==1.2.0
